refactor: move Postal::DKIMHeader to app/util/dkim_header
